tests refactoring out
