--- JOKES.md
@@ -1,9 +1,0 @@
-# Goose Jokes 🦢
-
-## Why did the Goose become a developer?
-
-Because it wanted to help debug all those "fowl" errors in the code!
-
----
-
-*Feel free to add more goose-related programming humor below!*
