+++ crates/goose-mcp/src/developer/mod.rs
@@ -5,3 +5,6 @@ mod shell;
 mod text_editor;
 
 pub mod rmcp_developer;
+
+#[cfg(test)]
+mod tests;
